requirements.txt
aiosmtplib==3.0.2
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
redis
APScheduler==3.11.0
bcrypt==4.0.1
blinker==1.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.23.0
cmdstanpy==1.3.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.3
coverage==7.10.7
cycler==0.12.1
cymem==2.0.11
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
fastapi-mail==1.5.0
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
gevent==25.9.1
greenlet==3.2.4
h11==0.16.0
holidays==0.83
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
loki-logger-handler==1.1.2
Mako==1.3.10
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
mypy==1.17.1
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.4
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathspec==0.12.1
pillow==12.0.0
pluggy==1.6.0
preshed==3.0.10
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
prophet==1.2.1
psycopg2-binary==2.9.10
pt_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.8.0/pt_core_news_sm-3.8.0-py3-none-any.whl#sha256=c304fa04db3af73cd08a250feacf560506e15a2ec2469bd1b09f06847f6b455c
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-logging-loki==0.3.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.11.3
requests==2.32.5
rfc3339==6.2
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
rsa==4.9.1
ruff==0.12.11
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
sentry-sdk==2.35.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.4.1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
srsly==2.5.1
stanio==0.5.1
starlette==0.47.3
sympy==1.14.0
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.1
typer==0.16.1
types-passlib==1.7.7.20250602
types-pyasn1==0.6.0.20250516
types-python-jose==3.5.0.20250531
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
Unidecode==1.4.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.1.0
weasel==0.4.1
websockets==15.0.1
wrapt==2.0.0
zope.event==6.0
zope.interface==8.0.1
